ci: fix the five real failures from the first blocking run
- fmt: repo rust-toolchain.toml pins a nightly without rustfmt;
  pin the job to stable via RUSTUP_TOOLCHAIN
- clippy: host-mint-tool -> rusb -> libudev-sys needs libudev-dev /
  libusb-1.0-0-dev on ubuntu; apt-install them; explicit
  RUSTUP_TOOLCHAIN so cargo matches the dtolnay-installed nightly
- adapter-test: the adapter spawns ./target/release/mint_server as a
  subprocess — build mint_server too (the original workflow never
  did; this is why adapter-test failed every run historically)
- deny: v2 action takes command + arguments; RUSTUP_TOOLCHAIN=stable
  (metadata only, no nightly needed)
- assets: TileAsset.data now &'static [u32] — the struct still said
  u16 after the array fix; this u16/u32 mismatch is what had been
  failing firmware-build (both workflows) all along, predating the
  dleq breakage

+++ firmware/src/boot_splash_assets.rs
@@ -11,7 +11,7 @@
 pub struct TileAsset {
     pub width: u16,
     pub height: u16,
-    pub data: &'static [u16],
+    pub data: &'static [u32],
 }
 
 /// Tile: tiny — 21x24 pixels (504 u32 ARGB values)
